Simplify leaderboard components
- Reduce main/runs/tracks leaderboard components into one all-encompassing leaderboard component that uses props to distinguish between all leaderboards
- Delete old leaderboard components
- Reduce leaderboard component even more by creating a leaderboard-navigation component that serves to provide page info and page functions that appears below the leaderboard
