+++ Trials.com/react-frontend/src/components/leaderboards/tracks-leaderboard-columns.tsx
@@ -1,7 +1,12 @@
+import { faSortAmountDown } from "@fortawesome/free-solid-svg-icons";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { Cell, Column, Row } from "react-table";
+import { infoTip } from "../help-info/info-tips";
+import { getTracksLB } from "../leaderboard-requests";
 import { SelectColumnFilter } from "./filters/select-filter";
 import { filterBetween, SliderColumnFilter } from "./filters/slider-filter";
 
-export const COLUMNS = [
+export const TracksLeaderboardColumns = [
     {
         Header: "Track", // Header is the title for the column being displayed (can be anything)
         accessor: "track_name", // accessor is based off of the prisma schemas name mapped for the database column you want
@@ -44,3 +49,61 @@ export const COLUMNS = [
         filter: "equals",
     },
 ];
+
+export const setTableBodyCell = (cell: Cell, row: Row) => {
+    if (cell.column.Header === "Track") {
+        return (
+            <a
+                href={
+                    "track/name=" +
+                    row.values.track_name +
+                    "&creator=" +
+                    row.values.creator
+                }
+            >
+                {cell.render("Cell")}
+            </a>
+        );
+    } else {
+        return <div>{cell.render("Cell")}</div>;
+    }
+};
+
+export const setTableHeaderInfoTip = (column: Column) => {
+    if (column.Header === "Ninja Points (Avg)") {
+        return infoTip(
+            "average-np-run",
+            "This column lists the ninja point averages of all runs on the given track"
+        );
+    } else if (column.Header === "Faults (Avg)") {
+        return infoTip(
+            "average-faults",
+            "This column lists the average faults for a run on the given track"
+        );
+    } else if (column.Header === "Consistency") {
+        return infoTip(
+            "consistency",
+            "This describes how consistent a track is, in other words how 'luck'-based it is"
+        );
+    } else {
+        return (
+            <span className="invisible-element">
+                <FontAwesomeIcon icon={faSortAmountDown} size="1x" />
+            </span>
+        );
+    }
+};
+
+export const resolveData = async (setData: any, date?: Date) => {
+    try {
+        const data = await getTracksLB(date ? date : undefined);
+
+        setData(data);
+    } catch (err) {
+        console.error(err);
+    }
+};
+
+export const tracksLBEffect = (setData: any, date?: Date, runs?: any) => {
+    resolveData(setData, date);
+};
