+++ Trials.com/react-frontend/src/pages/dashboard.tsx
@@ -1,5 +1,11 @@
 import { Chat } from "../components/General-Chat/chat";
-import { MainLeaderboard } from "../components/leaderboards/main-leaderboard";
+import { Leaderboard } from "../components/leaderboards/leaderboard";
+import {
+    MainLeaderboardColumns,
+    setTableBodyCell,
+    setTableHeaderInfoTip,
+    mainLBEffect,
+} from "../components/leaderboards/main-leaderboard-columns";
 import NavBar from "../components/navbar";
 
 export default function Dashboard(props: any) {
@@ -13,7 +19,13 @@ export default function Dashboard(props: any) {
                     <Chat />
                 </div>
                 <div className="dashboard-leaderboard">
-                    <MainLeaderboard />
+                    <Leaderboard
+                        columns={MainLeaderboardColumns}
+                        sortBy="total_ninja_points"
+                        setTableBodyCell={setTableBodyCell}
+                        setTableHeaderInfoTip={setTableHeaderInfoTip}
+                        effect={mainLBEffect}
+                    />
                 </div>
 
                 <div className="dashboard-space-element"></div>
