+++ Trials.com/react-frontend/src/pages/tracks.tsx
@@ -1,5 +1,11 @@
 import { Chat } from "../components/General-Chat/chat";
-import { TracksLeaderboard } from "../components/leaderboards/tracks-leaderboard";
+import { Leaderboard } from "../components/leaderboards/leaderboard";
+import {
+    TracksLeaderboardColumns,
+    setTableBodyCell,
+    setTableHeaderInfoTip,
+    tracksLBEffect,
+} from "../components/leaderboards/tracks-leaderboard-columns";
 import NavBar from "../components/navbar";
 
 export default function Tracks(props: any) {
@@ -13,7 +19,13 @@ export default function Tracks(props: any) {
                     <Chat />
                 </div>
                 <div className="tracks-leaderboard">
-                    <TracksLeaderboard />
+                    <Leaderboard
+                        columns={TracksLeaderboardColumns}
+                        effect={tracksLBEffect}
+                        sortBy="total_ninja_points"
+                        setTableBodyCell={setTableBodyCell}
+                        setTableHeaderInfoTip={setTableHeaderInfoTip}
+                    />
                 </div>
 
                 <div className="tracks-space-element"></div>
